test(agathos): fix test naming and imports for Agathos rename
- Rename test_argus.py -> test_agathos.py
- Fix imports: argus -> agathos package references
- Fix variable naming: self.argus -> self.agathos
- Update docstrings: ARGUS -> Agathos
- Fix compliance test paths and class references
- Fix provider_health test fixtures naming
- All 10 unit tests now pass
